Add debug icon when many targets on the same page
Normally, you notice the debugger itself and as such know that you’re in
debug mode. No need to waste more space on an icon then.

But when having multiple targets on the same page, you can’t know which
target the debugger is for. And there can be multiple debuggers on top
of each other. In that case an icon helps.