Fix error for `Browser.element` when used with a plain HTML file
You are supposed to be able to open a plain HTML file directly in your
browser, using the `file://` protocol. Only `Browser.application` does
not support that – Elm throws a runtime error.

We call `_Browser_getUrl` – if it exists – in `_Platform_initialize`,
because we need to remember the initial URL for `Browser.application`.

However, that’s the function that throws an error for `file://`. And it
exists even when using for example `Browser.element`.

This commit changes to specifically checking if we’re a
`Browser.application` instead, and only tries to call `_Browser_getUrl`
then.