Write entire error message to file, not just the bad JSON
I used elm-watch in a real project, and after a while I noticed an
`elm-watch-ElmMakeJsonParseError-xxx.json` file had been created. At
that point, the associated error message in the terminal was long gone.
It felt random to suddenly have a JSON file there.

This commit writes a .txt file instead, with the whole error in it to
make it easier to understand what it is.