Properly pause stdin on exit
Previously, we only called `logger.kill()` in one place, but there are
more places that lead to a program exit which also need to stop stdin
(otherwise Node.js waits for it).

This commit moves `logger.kill()` to `closeAll`, and starts using
`closeAll` when restarting, so that all places use `closeAll` and
properly close everything.

`logger.kill` was also renamed to `logger.pause` since that’s what it’s
actually doing, and to indicate that it can be resumed again after a
restart.