Replace whole rules since granular updates had edge cases
For example, `border: 4px solid black`. If you disabled that in the
devtools and then changed it to `border: 8px solid black` then you ended
up with 8px wide standard borders (non-solid, not black). Because the
only thing we truly changed was the width. But one expects the whole
property to be reflected, not just a part of it.