Fix improper kill if WebSocket already closed
This could happen:

1. We call `__ELM_WATCH.KILL_MATCHING` with the intention of killing a
   target.
2. Its WebSocket is not connected, but its not because the target is
   already killed. This skipped running the killing, assuming it was
   already killed.
3. The WebSocket reconnect timeout fires later, when another test is
   running. The TeaProgram’s internal `killed` variable is still `false`
   so it goes ahead and processes the message.
4. That causes a render, and happened to trigger `ON_RENDER` in the
   startup of another test, but too early, causing it to fail due to the
   elm-watch UI not existing yet.

This happened in the "typecheck only" test, affecting the "missing
window.Elm" test.

This commit changes `__ELM_WATCH.KILL_MATCHING` to _always_ properly
kill the target, regardless of whether the WebSocket is connected or
not.

This uncovered a timing problem in the "typecheck only" test. For now, I
added a sleep to make it pass.