Update snapshot for "Move UI" test
For commit 2de09a601d11bd3882e0a1e90022e04b2cf5a6d7.

That commit changes how code is eval:ed. Before:

    const f = new Function(code);
    f();

After:

    await import(`data:text/javascript,${encodeURIComponent(code)}`);

The new approach is slower, causing different timing. The end result is
the same, but the statuses in the snapshot change slightly (from one
valid sequence to another).

I updated the snapshots by temporarily changing the val implementation
to:

    await new Promise((resolve) => {
      setTimeout(resolve, 100);
    });
    const f = new Function(code);
    f();

That is the old implementation, but slowed down. I updated the
snapshots, and then reverted to the new implementation. The tests
passed. This shows that the new implementation is indeed slower, and
that the difference can be explained as a timing thing.