Validate open in editor command more
If the user sets the `ELM_WATCH_OPEN_EDITOR` but forgets to quote
the `file` variable as recommended in the docs, a malicious user
could send a web socket message with the file name set to a shell
command that then would be evalutated.

This commit validates that the file is imported by any target, and only
runs the command if it is. This prevents sending any shell command as
file name.

Backports commit b72ea3737541a6e57c00ebb6cb915b71f1da9bf8.

Conflicts:
	client/WebSocketMessages.ts
	src/Hot.ts
	tests/Hot.test.ts