Delay init cmds instead of reloading the page
Commit f73b857422ab5d3345cd08c822b4fa6d2d79011b introduced reloading of
the page in case init Cmds need to be re-run.

This commit reworks that in a smarter way: On page load, don’t run Cmds
from `init` immediately. Wait for the Web Socket to connect first. If we
get a hot reload, run the Cmds from that new `init` instead. Otherwise
proceed with the original (delayed) `init` Cmds.

This avoids an extra page reload, which turned out to be quite annoying.